# pkg/server/openapi.yaml
openapi: 3.1.0
info:
  title: native-ops daemon API
  version: set-at-runtime
  summary: Plan, deploy and inspect one Incus host from CI, an agent or a dashboard.
  description: |
    The native-ops daemon runs on the host it manages. Every change to the host is a **job** with a
    record; one change runs at a time (a second gets `409 busy`).

    **Deploys** are the normal way to change the host: push a protected `deploy-*` tag in the
    configuration repository and call `POST /v1/deploy` with it. The daemon reads that commit from the
    git server itself; nothing is uploaded, and the tag push is the approval. `POST /v1/deploy/plan`
    shows what a tag would change first; `GET /v1/deploys` says what the host runs now.

    **Roles** are ordered: viewer < planner < deployer < admin. Each operation names the least role it
    needs in `x-native-ops-role`. A token with a **scope** (for a fleet manager) may only call the
    operations marked `x-native-ops-scoped: true`, and only for instances its scope allows.

    Some operations exist only when the daemon is started with the flag in `x-native-ops-enabled-by`;
    otherwise they answer 404.

    Errors are `{"error": "<message>", "code": "<code>"}`. A failure to read the host is a generic
    `502 host_unavailable` that never echoes internal text.

    The same operations are available to AI agents as MCP tools at `POST /mcp` (see docs/api.md).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: "{daemon}"
...
security:
  - bearer: []
...
paths:
...
  /v1/whoami:
    get:
      operationId: whoami
      tags: [meta]
      summary: Which token and role you are
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      responses:
        "200":
          description: The caller.
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  role: { $ref: "#/components/schemas/Role" }
                  version: { type: string }
                  scope: { $ref: "#/components/schemas/Scope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/status:
    get:
      operationId: status
      tags: [host]
      summary: The host's state
      description: |
        Instances, data volumes, images, host metrics (load, memory, swap, pool disk, uptime) and
        warnings; with the edge and DNS configured, the edge's routes and certificates and the DNS
        records. Instance config is reported by key name only. Cached for a few seconds.
      x-native-ops-role: viewer
      responses:
        "200": { description: "The snapshot.", content: { application/json: { schema: { type: object } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/HostUnavailable" }
  /v1/deploy:
    post:
      operationId: deploy
      tags: [deploys]
      summary: Deploy a protected tag
      description: |
        Starts a deploy job for the commit the tag points at. The job:

        1. resolves the tag on the git server and checks a protection rule covers it (otherwise it fails
           without fetching anything);
        2. downloads that commit's tree from the git server;
        3. if fleet.yml pins another daemon release, upgrades to it, restarts, and resumes the deploy as
           a new job on the new binary (this job ends with result `daemon_upgraded`);
        4. plans it; a blocked plan fails the job and nothing changes; nothing to change ends it with
           result `no_changes`;
        5. applies the plan (result `applied`).

        Follow the job with `GET /v1/jobs/{id}?wait=60`, or see it in `GET /v1/deploys`.
      x-native-ops-role: deployer
      x-native-ops-enabled-by: --enable-apply, --git-url and --deploy-repo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagRequest" }
      responses:
        "202":
          description: The deploy job started. Location names the job.
          headers:
            Location: { schema: { type: string, examples: [/v1/jobs/j-1791054271-18f384d8] } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  job: { $ref: "#/components/schemas/Job" }
                  tag: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Busy" }
  /v1/deploy/plan:
    post:
      operationId: planDeploy
      tags: [deploys]
      summary: What deploying a tag would change
      description: |
        Resolves the tag, reads its commit from the git server and plans it against the host now,
        without changing anything and without starting a job. The plan is recorded (`GET /v1/plans`).

        An unprotected tag is still planned, with `deployable: false` and the reason. When fleet.yml pins
        a daemon release other than the running one, `daemon.upgrade` is true: the deploy would upgrade
        first, and the pinned release makes the plan that is applied.
      x-native-ops-role: planner
      x-native-ops-enabled-by: --enable-apply, --git-url and --deploy-repo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagRequest" }
      responses:
        "200":
          description: The plan.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeployPlan" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: The git server could not resolve the tag (no such tag, or not a deploy tag).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { description: "Too many plans are running; retry after the Retry-After header." }
        "502":
          description: The commit could not be read from the git server (git_unavailable), or the host could not be read (host_unavailable).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/deploys:
    get:
      operationId: listDeploys
      tags: [deploys]
      summary: What the host runs, and the deploy history
      description: |
        `current` is the newest deploy that left the host matching its tag (result `applied` or
        `no_changes`). Changes made since by other means (an apply, an instance change, an image build)
        are in `GET /v1/jobs`. `running` is the deploy in progress, if any.
      x-native-ops-role: viewer
      x-native-ops-enabled-by: --enable-apply, --git-url and --deploy-repo
      parameters:
        - { name: limit, in: query, description: "How many deploys to list (default 50).", schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: The deploys.
          content:
            application/json:
              schema:
                type: object
                required: [current, running, deploys]
                properties:
                  current: { oneOf: [{ $ref: "#/components/schemas/Deploy" }, { type: "null" }] }
                  running: { oneOf: [{ $ref: "#/components/schemas/Deploy" }, { type: "null" }] }
                  deploys: { type: array, items: { $ref: "#/components/schemas/Deploy" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Jobs, newest first (without logs)
      description: A scoped token sees only jobs it started and jobs about instances its scope allows.
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      parameters:
        - name: kind
          in: query
          description: A job kind, or a family of kinds ending in a colon (`instance:`).
          schema: { type: string, examples: [deploy, apply, "instance:", image:build, image:prune, edge:apply, backup, restore, dns:sync, daemon:upgrade] }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/JobStatus" } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: The jobs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs: { type: array, items: { $ref: "#/components/schemas/Job" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: The plans that were made, and where each stands
      x-native-ops-role: viewer
      responses:
        "200": { description: "The plans (pending, approved, used, expired, stale, blocked).", content: { application/json: {} } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/instances:
    get:
      operationId: listInstances
      tags: [instances]
      summary: The tenant instances the token may see
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      x-native-ops-enabled-by: --enable-instances
      responses:
        "200": { description: "The instances.", content: { application/json: {} } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/images/prune:
    post:
      operationId: pruneImages
      tags: [images]
      summary: Image retention, as a job
      description: Deletes images no instance runs that are orphans or tags older than the newest --image-keep per app. The job's log lists each.
      x-native-ops-role: deployer
      x-native-ops-enabled-by: --enable-image-build
      parameters:
        - { name: dry_run, in: query, description: "Only report.", schema: { type: string, enum: ["1", "true"] } }
      responses:
        "202": { $ref: "#/components/responses/JobStarted" }
        "409": { $ref: "#/components/responses/Busy" }
...
  /v1/backups/restore:
    post:
      operationId: restore
      tags: [maintenance]
      summary: Restore a volume in place (keeping a copy) or as a new volume
      x-native-ops-role: admin
      x-native-ops-enabled-by: --enable-backup
      parameters:
        - { name: volume, in: query, required: true, schema: { type: string } }
        - { name: from, in: query, description: "The object key, or latest.", schema: { type: string, default: latest } }
        - { name: as, in: query, description: "Restore under this new volume name.", schema: { type: string } }
        - { name: force, in: query, description: "Stop the containers that mount the volume.", schema: { type: string, enum: ["1", "true"] } }
      requestBody: { $ref: "#/components/requestBodies/Tree" }
      responses:
        "202": { $ref: "#/components/responses/JobStarted" }
        "409": { $ref: "#/components/responses/Busy" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: An API token (nops_...), created by an admin (POST /v1/tokens or native-ops token create). The UI's session cookie also works, except at /mcp.
...
  responses:
...
    BadRequest:
      description: The request is malformed (bad_request, bad_config).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: No valid token (unauthorized). WWW-Authenticate says how to authenticate.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: The token's role (or scope) cannot do that (forbidden).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Busy:
      description: Another change is running on this host (busy); the message names its job. Wait for it and retry.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    HostUnavailable:
      description: The host could not be read (host_unavailable).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: "What went wrong, for a person." }
        code:
          type: string
          description: Stable, for a program.
          examples: [bad_request, bad_config, unauthorized, forbidden, not_found, busy, tag_unresolved, git_unavailable, host_unavailable, plan_changed, blocked, stale, not_approvable, not_approved, approval_used, approval_expired, not_managed, internal]
    Role:
      type: string
      enum: [viewer, planner, deployer, admin]
    Scope:
      type: object
      description: Limits a token to tenant instances (names, images and domains, as globs).
      properties:
        names: { type: array, items: { type: string } }
        images: { type: array, items: { type: string } }
        domains: { type: array, items: { type: string } }
    TagRequest:
      type: object
      required: [tag]
      additionalProperties: false
      properties:
        tag: { type: string, description: "A deploy tag pushed in the configuration repository.", examples: [deploy-2026.10.03] }
    JobStatus:
      type: string
      enum: [running, succeeded, failed, interrupted]
      description: interrupted means the daemon stopped while the job ran; run it again (applies converge).
    Job:
      type: object
      required: [id, kind, status, actor, created]
      properties:
        id: { type: string, examples: [j-1791054271-18f384d8] }
        kind: { type: string, examples: [deploy, apply, "instance:put", image:build] }
        status: { $ref: "#/components/schemas/JobStatus" }
        actor: { type: string, description: "The token or user that started it." }
        sha: { type: string, description: "The commit (a deploy's, once resolved)." }
        service: { type: string, description: "What the job is about; for a deploy, the tag." }
        plan_hash: { type: string }
        created: { type: string, format: date-time }
        finished: { type: string, format: date-time }
        error: { type: string }
        result: { $ref: "#/components/schemas/DeployResult" }
        log: { type: string, description: "For deployers and admins; not in lists." }
        log_truncated: { type: boolean }
...
    DeployPlan:
      type: object
      required: [tag, sha, deployable, daemon, hash, exit, counts, plan, text]
      properties:
        tag: { type: string }
        sha: { type: string }
        deployable: { type: boolean, description: "Whether POST /v1/deploy would accept the tag (a protection rule covers it)." }
        protected_by: { type: string, description: "The protection rule, when deployable." }
        reason: { type: string, description: "Why not, when not deployable." }
        daemon:
          type: object
          properties:
            running: { type: string }
            pinned: { type: string }
            upgrade: { type: boolean, description: "The deploy would upgrade the daemon first." }
        hash: { type: string }
        exit: { $ref: "#/components/schemas/PlanExit" }
        counts: { $ref: "#/components/schemas/PlanCounts" }
        plan: { $ref: "#/components/schemas/FleetPlan" }
        text: { type: string, description: "The plan, readable." }
    PlanExit:
      type: integer
      enum: [0, 1, 2]
      description: 0 nothing to change, 1 blocked (apply would fail), 2 changes pending.
...